RPCT-2312, u7: Work on warnings from static code analyzer

+++ u7/CentralWidget.cpp
@@ -7,7 +7,7 @@ CentralWidget::CentralWidget(QWidget* parent) :
 	QTabWidget(parent)
 {
 	QSize sz = fontMetrics().size(Qt::TextSingleLine, "APPLICATION LOGIC");
-	sz.setHeight(sz.height() * 1.75);
+	sz.setHeight(static_cast<int>(sz.height() * 1.75));
 
 	QString ss = QString("QTabBar::tab{ min-width: %1px; min-height: %2px;}").arg(sz.width()).arg(sz.height());
 	tabBar()->setStyleSheet(ss);
