+++ u7/CheckInDialog.cpp
@@ -70,7 +70,10 @@ void CheckInDialog::showEvent(QShowEvent*)
 	// Resize depends on monitor size, DPI, resolution
 	//
 	QRect screen = QDesktopWidget().availableGeometry(parentWidget());
-	resize(screen.width() * 0.30, screen.height() * 0.30);
+
+	resize(static_cast<int>(screen.width() * 0.30),
+		   static_cast<int>(screen.height() * 0.30));
+
 	move(screen.center() - rect().center());
 
 	return;
